requirements.txt
aiohttp==3.6.2
APScheduler==3.6.1
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
defusedxml==0.6.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
jira==2.0.0
MarkupSafe==1.1.1
multidict==4.5.2
oauthlib==3.1.0
pbr==5.4.3
pycparser==2.19
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
six==1.12.0
slackclient==2.5.0
tzlocal==2.0.0
urllib3==1.25.6
Werkzeug==1.0.0
yarl==1.3.0
